2040: separate the ECAP functionality from the ECA (eca.go -> ecap.go)
